requirements.txt
PyOpenGL
PyOpenGL-accelerate
pygame
